+++ source/ftrack_api/session.py
@@ -12,6 +12,7 @@ from builtins import object
 import json
 import logging
 import collections
+from six.moves import collections_abc
 import datetime
 import os
 import getpass
@@ -264,6 +265,11 @@ class Session(object):
         if auto_connect_event_hub:
             # Connect to event hub in background thread so as not to block main
             # session usage waiting for event hub connection.
+
+            # set the connection as initialising from the main thread so that
+            # we can queue up any potential published messages.
+            self._event_hub.init_connection()
+
             self._auto_connect_event_hub_thread = threading.Thread(
                 target=self._event_hub.connect
             )
@@ -1812,7 +1818,7 @@ class Session(object):
 
     def _decode(self, item):
         '''Return *item* transformed into appropriate representation.'''
-        if isinstance(item, collections.Mapping):
+        if isinstance(item, collections_abc.Mapping):
             if '__type__' in item:
                 if item['__type__'] == 'datetime':
                     item = arrow.get(item['value'])
@@ -2463,7 +2469,7 @@ class OperationRecordingContext(object):
         self._session.record_operations = self._current_record_operations
 
 
-class OperationPayload(collections.MutableMapping):
+class OperationPayload(collections_abc.MutableMapping):
     '''Represent operation payload.'''
 
     def __init__(self, *args, **kwargs):